+++ content/src/main/kotlin/Content.kt
@@ -1,5 +1,6 @@
 import org.w3c.dom.url.URL
 import websites.AlcCoJp
+import websites.TranslateGoogleCom
 import websites.WeblioJp
 import kotlin.browser.window
 
@@ -7,6 +8,7 @@ external val chrome: dynamic
 
 fun main() {
     when (URL(window.location.href).hostname) {
+        "translate.google.com" -> TranslateGoogleCom().register()
         "eow.alc.co.jp" -> AlcCoJp().register()
         "ejje.weblio.jp" -> WeblioJp().register()
     }